+++ src/Entity/IdBasedEntity.php
@@ -13,26 +13,27 @@ declare(strict_types=1);
 
 namespace EightMarq\CoreBundle\Entity;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use EightMarq\CoreBundle\Entity\Interfaces\Behavior\SoftDeleteableInterface;
+use EightMarq\CoreBundle\Entity\Interfaces\Behavior\TimestampableInterface;
 use EightMarq\CoreBundle\Entity\Interfaces\IdBasedEntityInterface;
+use EightMarq\CoreBundle\Entity\Traits\Behavior\SoftDeleteable\SoftDeleteableTrait;
+use EightMarq\CoreBundle\Entity\Traits\Behavior\Timestampable\TimestampableTrait;
 use EightMarq\CoreBundle\Entity\Traits\IdBasedEntityMethodsTrait;
 use Knp\DoctrineBehaviors\Contract\Entity\BlameableInterface;
-use Knp\DoctrineBehaviors\Contract\Entity\SoftDeletableInterface;
-use Knp\DoctrineBehaviors\Contract\Entity\TimestampableInterface;
 use Knp\DoctrineBehaviors\Model\Blameable\BlameableTrait;
-use Knp\DoctrineBehaviors\Model\SoftDeletable\SoftDeletableTrait;
-use Knp\DoctrineBehaviors\Model\Timestampable\TimestampableTrait;
 
 #[ORM\MappedSuperclass]
-class IdBasedEntity implements IdBasedEntityInterface, TimestampableInterface, BlameableInterface, SoftDeletableInterface
+class IdBasedEntity implements IdBasedEntityInterface, TimestampableInterface, BlameableInterface, SoftDeleteableInterface
 {
     use IdBasedEntityMethodsTrait;
     use BlameableTrait;
-    use SoftDeletableTrait;
+    use SoftDeleteableTrait;
     use TimestampableTrait;
 
     #[ORM\Id]
-    #[ORM\Column(type: 'integer')]
+    #[ORM\Column(type: Types::INTEGER)]
     #[ORM\GeneratedValue(strategy: 'AUTO')]
     protected int|null $id = null;
 }