+++ src/Entity/Traits/Behavior/SoftDeleteable/SoftDeleteablePropertiesTrait.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * This file is part of the EightMarq Symfony bundles.
+ *
+ * (c) Norbert Schvoy <norbert.schvoy@eightmarq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EightMarq\CoreBundle\Entity\Traits\Behavior\SoftDeleteable;
+
+use DateTime;
+use Doctrine\DBAL\Types\Types;
+use Doctrine\ORM\Mapping as ORM;
+
+trait SoftDeleteablePropertiesTrait
+{
+    #[ORM\Column(type: Types::DATETIME_MUTABLE, nullable: true)]
+    protected DateTime|null $deletedAt = null;
+}