+++ src/EventListener/TimestampableEventListener.php
@@ -1,0 +1,44 @@
+<?php
+
+/**
+ * This file is part of the EightMarq Symfony bundles.
+ *
+ * (c) Norbert Schvoy <norbert.schvoy@eightmarq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EightMarq\CoreBundle\EventListener;
+
+use Doctrine\Bundle\DoctrineBundle\Attribute\AsDoctrineListener;
+use Doctrine\ORM\Event\LoadClassMetadataEventArgs;
+use Doctrine\ORM\Events;
+use EightMarq\CoreBundle\Entity\Interfaces\Behavior\TimestampableInterface;
+
+#[AsDoctrineListener(event: Events::loadClassMetadata)]
+final class TimestampableEventListener
+{
+    public function loadClassMetadata(LoadClassMetadataEventArgs $loadClassMetadataEventArgs): void
+    {
+        $classMetadata = $loadClassMetadataEventArgs->getClassMetadata();
+
+        if (null === $classMetadata->reflClass) {
+            // Class has not yet been fully built, ignore this event
+            return;
+        }
+
+        if (!is_a($classMetadata->reflClass->getName(), TimestampableInterface::class, true)) {
+            return;
+        }
+
+        if ($classMetadata->isMappedSuperclass) {
+            return;
+        }
+
+        $classMetadata->addLifecycleCallback('updateDateTimes', Events::prePersist);
+        $classMetadata->addLifecycleCallback('updateDateTimes', Events::preUpdate);
+    }
+}