+++ src/Services/DateTimeProvider.php
@@ -1,0 +1,28 @@
+<?php
+
+/**
+ * This file is part of the EightMarq Symfony bundles.
+ *
+ * (c) Norbert Schvoy <norbert.schvoy@eightmarq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EightMarq\CoreBundle\Services;
+
+use DateTime;
+use DateTimeZone;
+
+class DateTimeProvider
+{
+    public static function getCurrent(): DateTime
+    {
+        $currentDate = new DateTime();
+        $currentDate->setTimezone(new DateTimeZone(date_default_timezone_get()));
+
+        return $currentDate;
+    }
+}