+++ tests/Fixtures/Entity/Behavior/TimestampableEntity.php
@@ -1,0 +1,36 @@
+<?php
+
+/**
+ * This file is part of the EightMarq Symfony bundles.
+ *
+ * (c) Norbert Schvoy <norbert.schvoy@eightmarq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EightMarq\CoreBundle\Tests\Fixtures\Entity\Behavior;
+
+use Doctrine\DBAL\Types\Types;
+use Doctrine\ORM\Mapping as ORM;
+use EightMarq\CoreBundle\Entity\IdBasedEntity;
+
+#[ORM\Entity]
+#[ORM\Table]
+class TimestampableEntity extends IdBasedEntity
+{
+    #[ORM\Column(type: Types::STRING)]
+    private string $value;
+
+    public function getValue(): string
+    {
+        return $this->value;
+    }
+
+    public function setValue(string $value): void
+    {
+        $this->value = $value;
+    }
+}